fix type signedness warning

+++ libs/ledger/include/ledger/chain/main_chain_miner.hpp
@@ -97,8 +97,8 @@ private:
   Timestamp CalculateNextBlockTime(T &rng)
   {
     auto jitterrange = block_interval_ / 10;
-    unsigned long jitterrange_us =
-      std::abs(std::chrono::duration_cast<std::chrono::microseconds>(jitterrange).count());
+    unsigned long jitterrange_us = static_cast<unsigned long>(std::abs(
+        std::chrono::duration_cast<std::chrono::microseconds>(jitterrange).count()));
     auto random_us = rng() % jitterrange_us - (jitterrange_us / 2);
 
     Timestamp block_time = Clock::now() + block_interval_ + Microseconds{random_us};
